requirements.txt
Flask==3.0.3
flask-cors==4.0.1
psycopg2-binary